requirements.txt
cffi==1.15.1
inflection==0.5.1
pycparser==2.21
raylib==4.2.1.1
requests==2.28.1
Pillow==9.2.0